dom/tests/mochitest/bugs/test_bug384122.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=384122
-->
<head>
  <title>Test for Bug 384122</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=384122">Mozilla Bug 384122</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 384122 **/

var foundQI = false;
for (var property in Node)
  if (property == "QueryInterface")
    foundQI = true;
is(foundQI, false,
   "QueryInterface should not be enumerable on DOM interface objects");

</script>
</pre>
</body>
</html>
